root

+++ lib/builders/genericPack.nix
@@ -8,18 +8,23 @@
 , packList
 }:
 
-stdenv.mkDerivation {
+stdenv.mkDerivation rec {
   pname = "${name}-bundle";
   inherit version;
   srcs = packList;
   nativeBuildInputs = [
     gnutar
   ];
+  sourceRoot = pname;
   buildPhase = ''
-    for pack in $(ls); do
-      mv "$pack/*" .
-      mv "$pack/.*" .
-      rmdir "$pack"
+    for pack in $(ls ..); do
+      if [[ "$pack" -eq "${pname}" ]]; then
+        true
+      else
+        mv "$pack/*" .
+        mv "$pack/.*" .
+        rmdir "$pack"
+      fi
     done
     tar cvhzf ./out.tar.gz --no-same-permissions --no-same-owner ./*
   '';
